requirements.txt
flask==2.3.2
werkzeug==2.3.6
easyocr==1.6.2
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
spacy==3.5.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
python-dotenv==1.0.0
pymongo 